pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usda_fdc"
version = "0.1.4"
description = "A comprehensive Python library for interacting with the USDA Food Data Central API"
readme = "README.md"
authors = [{ name = "Michael McGarrah", email = "mcgarrah@gmail.com" }]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = ["requests>=2.25.0", "pint>=0.17", "python-dotenv>=0.19.0"]

[project.optional-dependencies]
django = ["Django>=3.2"]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "black>=21.5b2",
    "isort>=5.9.1",
    "mypy>=0.812",
    "flake8>=3.9.2",
]
docs = ["sphinx>=4.0.2", "sphinx-rtd-theme>=0.5.2"]

[project.urls]
Homepage = "https://github.com/mcgarrah/usda_fdc_python"
Documentation = "https://usda-fdc.readthedocs.io"
Issues = "https://github.com/mcgarrah/usda_fdc_python/issues"

[project.scripts]
fdc = "usda_fdc.cli:main"

[tool.setuptools]
packages = ["usda_fdc"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
